requirements.txt
asgiref==3.8.1
razorpay==1.4.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.22.0
idna==3.10
inflection==0.5.1
packaging==24.1
pillow==11.1.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.2.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
